added window.confirm to delete button, fixed study button link

+++ src/Layout/Home/DeckView.js
@@ -2,8 +2,11 @@ import React from "react";
 
 function DeckView({deck, deleteDeck }) {
     function deleteButtonHandler() {
-        deleteDeck()
-        window.location.reload(false)
+        const confirm = window.confirm("Delete this deck? You will not be able to recover it.")
+        if (confirm) {
+            deleteDeck()
+            window.location.reload(false)
+        }
     }
     return (
         <>
@@ -12,7 +15,7 @@ function DeckView({deck, deleteDeck }) {
             <h5 className="card-title">{deck.name}</h5>
             <p className="card-text">{deck.description}</p>
             <a href="/view" className="card-link"><button>View</button></a>
-            <a href="/study" className="card-link"><button>Study</button></a>
+            <a href={`/decks/${deck.id}/study`} className="card-link"><button>Study</button></a>
             <a className="card-link"><button onClick={deleteButtonHandler}>Delete</button></a>
             </div>
         </div>
